requirements.txt
crewai
crewai_tools
openai
litellm
python-dotenv
pyyaml
loguru
rich
requests
tqdm
